+++ src/main/java/spot/Ui.java
@@ -1,6 +1,9 @@
+package spot;
+
 import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
 import java.util.Scanner;
+
+import spot.task.Task;
 
 public class Ui {
