+++ src/main/java/spot/command/AddToDoCommand.java
@@ -1,3 +1,11 @@
+package spot.command;
+
+import spot.exception.SpotException;
+import spot.Storage;
+import spot.TaskList;
+import spot.Ui;
+import spot.task.ToDo;
+
 public class AddToDoCommand extends Command {
     private String description;
