+++ src/main/java/spot/command/ListCommand.java
@@ -1,3 +1,9 @@
+package spot.command;
+
+import spot.Storage;
+import spot.TaskList;
+import spot.Ui;
+
 public class ListCommand extends Command {
 
     @Override
