+++ src/main/java/spot/command/MarkCommand.java
@@ -1,3 +1,10 @@
+package spot.command;
+
+import spot.exception.SpotException;
+import spot.Storage;
+import spot.TaskList;
+import spot.Ui;
+
 public class MarkCommand extends Command {
 
     private int position;
