+++ src/main/java/spot/command/UnmarkCommand.java
@@ -1,14 +1,21 @@
-public class DeleteCommand extends Command {
+package spot.command;
+
+import spot.exception.SpotException;
+import spot.Storage;
+import spot.TaskList;
+import spot.Ui;
+
+public class UnmarkCommand extends Command {
 
     private int position;
 
-    public DeleteCommand(int position) {
+    public UnmarkCommand(int position) {
         this.position = position;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws SpotException {
-        tasks.deleteTask(position);
+        tasks.unmarkTask(position);
     }
 
     @Override
