+++ src/main/java/spot/exception/SpotException.java
@@ -1,3 +1,5 @@
+package spot.exception;
+
 public class SpotException extends Exception{
     public SpotException(String message) {
         super(message);
