+++ src/main/java/spot/task/Event.java
@@ -1,3 +1,5 @@
+package spot.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
