+++ src/main/java/spot/task/Task.java
@@ -1,3 +1,5 @@
+package spot.task;
+
 import java.time.LocalDate;
 
 public abstract class Task {
